Delete the duplicate sentence!
Complex models with high accuracy usually require a lot of memory, so there will be limitations if they are deployed on resource-constrained devices. Complex models with high accuracy usually require a lot of memory, so there will be limitations if they are deployed on resource-constrained devices. ---->> Complex models with high accuracy usually require a lot of memory, so there will be limitations if they are deployed on resource-constrained devices.